server/pyproject.toml
[project]
name = "chorus-server"
version = "0.1.0"
description = ""
authors = [
    {name = "Maanas Arora",email = "maanasarora23@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy (>=2.3.1,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "fastapi (>=0.115.14,<0.116.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "scikit-learn (>=1.7.0,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pytest (>=8.0.0,<9.0.0)",
    "httpx (>=0.27.0,<1.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "chorus-engine @ engine-dist/chorus_engine-0.1.1-py3-none-any.whl"
]


[tool.poetry]
package-mode = false


[tool.pytest.ini_options]
pythonpath = ["."]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
